+++ nuttx/configs/spark/src/stm32_cxxinitialize.c
@@ -1,5 +1,5 @@
 /************************************************************************************
- * configs/spark/src/up_cxxinitialize.c
+ * configs/spark/src/stm32_cxxinitialize.c
  *
  *   Copyright (C) 2013 Gregory Nutt. All rights reserved.
  *   Author: Gregory Nutt <gnutt@nuttx.org>
@@ -49,7 +49,7 @@
 #if defined(CONFIG_HAVE_CXX) && defined(CONFIG_HAVE_CXXINITIALIZE)
 
 /************************************************************************************
- * Definitions
+ * Pre-processor Definitions
  ************************************************************************************/
 /* Debug ****************************************************************************/
 /* Non-standard debug that may be enabled just for testing the static constructors */
