+++ nuttx/configs/spark/src/stm32_userleds.c
@@ -1,5 +1,5 @@
 /****************************************************************************
- * configs/spark/src/up_leds.c
+ * configs/spark/src/stm32_userleds.c
  *
  *   Copyright (C) 2011, 2013 Gregory Nutt. All rights reserved.
  *   Author: Gregory Nutt <gnutt@nuttx.org>
@@ -55,7 +55,7 @@
 #ifndef CONFIG_ARCH_LEDS
 
 /****************************************************************************
- * Definitions
+ * Pre-processor Definitions
  ****************************************************************************/
 
 /* CONFIG_DEBUG_LEDS enables debug output from this file (needs CONFIG_DEBUG
